Generate square puzzle impressions
Now generating a separate set of square impresssions which I'm also
providing to the overlay.

My first thought was to implement some basic state management so the
overlay can handle both sets of impresssions and swap between them,
but I'm now thinknig it might be better just to keep the overlay logic
as-is and only have it handle a single set, while the switching can be
handled from outside.

Maybe the overlay instance can simply expose a method for updating its
impression set?