Puzzle impressions journey working well
- Puzzle shape field now working and allows the user to select
  rectangular of square puzzle shapes
- When selected, the puzzle impression updates to the first available
  config for the given shape, and the size slider updates to reflect the
new configs
- Selecting puzzle sizes with the slider updates the impression,
and the change from one size to the next looks fast and clean.

TODO: There is a very slight but noticeable pause in the impression
redraw when the shape is changed. This is because I'm removing one large
set of elements with another. Maybe look in to improving this?