WIP: Puzzle pieces beginning to render...
I'm getting closer to a successful render, but there is a disconnect
between the original image size and the size of the puzzle we generate.

I need to think about the space available for solving, and what size the
puzzle pieces should be based on this.

The first step is to determine the minimum size for the play area, and
how much space should be given to the out-of-bounds area, and then work
out the puzzle size from here.

Below this minimum size, I should instead provide a cosmetic border, say
10px, and forego an out-of-bounds space entirely if there is no space
for one.