wip: Getting there...
I now have a rectangular puzzle solving area rendering.

I'm now going with the approach of setting the solving area to be
an arbitrary size, maintaining the aspect ratio of the puzzle image,
and setting the piece size accordingly.

Doing it this way should give us more flexibility and enable us to use
the fullsize image at all times, regardless of the size of the viewport.
It will also allow us to maximise the image quality presented to the
user because the pieces will always be rendered at a size that is
proportional to the viewport.