Big achievement: SVG puzzle pieces now rendering correctly
Finally figured out what was causing the puzzle pieces to be rendered to
the wrong size / proportions:

While experimenting with the piece types and the getJigsawShapeSvgString
method, I was passing the wrong value for 'piece size' in to the
function, and this was causing the path to be drawn bigger than it
should be.

Now we can move forward and update the puzzle solving logic to
accomodate our new pieces.