Initial persistence fixed
- Puzzle pieces are now being saved correctly when the puzzle is first
  created, and their position is saved after 'arrange around edges' is
called.

The existing save data was ignored because the puzzle's lastSaveDate was
not being set on the server side when the pieces updated.

This was caused by trying to get the puzzleId from the wrong place as, when
updating a collection of pieces, the logic was treating the collection
as an object with a puzzleId member instead of extracting it from the
first object in the collection.