wip: Get connector bounding boxes from path strings
I'm consistently finding that the way I've been calculating
bounding boxes for connectors is awkward and innaccurate,
so I'm revisiting the idea of extracting them from the
SVG path strings instead.

This should be a much cleaner and more elegant solution,
and it should decouple the existing logic and allow for more flexibility
and less code as we move toward natural and wild piece shapes.

Also some persistence work.