checkpoint: Grouping
- Container for new groups now appears to be correct

Now starting to think about how to make grouping work for svgs.
We need a clip path that describes only the perimiter of the pieces in
the group. We also need to accomodate gaps in the group where a piece is
missing between other grouped pieces.

I'm thinking of splitting the perimeter into rows because this should
allow us to leave gaps, and this might give us a performance boost
if we only have to redraw the effected paths in the svg when a piece is
added, or two groups are merged.